+++ tests/assets/mpii_json_dataset/dataset_with_numpy_files/000000003.jpg
