+++ tests/assets/mpii_json_dataset/dataset_with_numpy_files/jnt_visible.npy
