+++ tests/assets/mpii_json_dataset/dataset_with_numpy_files/mpii_pos_gt.npy
