+++ tests/assets/mpii_json_dataset/dataset_wo_numpy_files/mpii_annotations.json
